+++ prompt.py
@@ -942,8 +942,8 @@ def main():
     group = parser.add_mutually_exclusive_group()
     group.add_argument("-m", "--mainnet", action="store_true", default=False,
                        help="Use MainNet instead of the default TestNet")
-    group.add_argument("-p", "--privnet", action="store_true", default=False,
-                       help="Use PrivNet instead of the default TestNet")
+    group.add_argument("-p", "--privnet", nargs="?", metavar="host", const=True, default=False,
+                       help="Use a private net instead of the default TestNet, optionally using a custom host (default: 127.0.0.1)")
     group.add_argument("--coznet", action="store_true", default=False,
                        help="Use the CoZ network instead of the default TestNet")
     group.add_argument("-c", "--config", action="store", help="Use a specific config file")
@@ -970,7 +970,7 @@ def main():
         settings.setup_mainnet()
     elif args.privnet:
         try:
-            settings.setup_privnet()
+            settings.setup_privnet(args.privnet)
         except PrivnetConnectionError as e:
             logger.error(str(e))
             return